+++ packages/rich-text-html-renderer/src/__test__/index.test.ts
@@ -1,5 +1,5 @@
 import cloneDeep from 'lodash/cloneDeep';
-import { Block, BLOCKS, Document, INLINES, MARKS } from '@contentful/rich-text-types';
+import { Block, BLOCKS, Document, INLINES, MARKS, ResourceLink } from '@contentful/rich-text-types';
 import { documentToHtmlString, Options } from '../index';
 import {
   embeddedEntryDoc,
@@ -15,6 +15,7 @@ import {
   quoteDoc,
   ulDoc,
   tableWithHeaderDoc,
+  embeddedResourceDoc,
 } from './documents';
 import inlineEntity from './documents/inline-entity';
 
@@ -42,6 +43,22 @@ describe('documentToHtmlString', () => {
       {
         doc: headingDoc(BLOCKS.HEADING_2),
         expected: '<h2>hello world</h2>',
+      },
+      {
+        doc: headingDoc(BLOCKS.HEADING_3),
+        expected: '<h3>hello world</h3>',
+      },
+      {
+        doc: headingDoc(BLOCKS.HEADING_4),
+        expected: '<h4>hello world</h4>',
+      },
+      {
+        doc: headingDoc(BLOCKS.HEADING_5),
+        expected: '<h5>hello world</h5>',
+      },
+      {
+        doc: headingDoc(BLOCKS.HEADING_6),
+        expected: '<h6>hello world</h6>',
       },
     ];
 
@@ -183,6 +200,20 @@ describe('documentToHtmlString', () => {
     expect(documentToHtmlString(document)).toEqual(expected);
   });
 
+  it('renders default resource link block', () => {
+    const resourceLink: ResourceLink = {
+      sys: {
+        urn: 'crn:contentful:::content:spaces/6fqi4ljzyr0e/environments/master/entries/9mpxT4zsRi6Iwukey8KeM',
+        type: 'ResourceLink',
+        linkType: 'Contentful:Entry',
+      },
+    };
+    const document: Document = embeddedResourceDoc(resourceLink);
+    const expected = `<div></div>`;
+
+    expect(documentToHtmlString(document)).toEqual(expected);
+  });
+
   it('renders ordered lists', () => {
     const document: Document = olDoc;
     const expected = `<ol><li><p>Hello</p></li><li><p>world</p></li></ol><p></p>`;
@@ -267,8 +298,8 @@ describe('documentToHtmlString', () => {
       target: {
         sys: {
           id: '9mpxT4zsRi6Iwukey8KeM',
-          link: 'Link',
-          type: 'Asset',
+          type: 'Link',
+          linkType: 'Asset',
         },
       },
     };
@@ -283,8 +314,8 @@ describe('documentToHtmlString', () => {
       target: {
         sys: {
           id: '9mpxT4zsRi6Iwukey8KeM',
-          link: 'Link',
-          type: 'Entry',
+          type: 'Link',
+          linkType: 'Entry',
         },
       },
     };
@@ -294,18 +325,50 @@ describe('documentToHtmlString', () => {
     expect(documentToHtmlString(document)).toEqual(expected);
   });
 
+  it('renders resource hyperlink', () => {
+    const entry = {
+      target: {
+        sys: {
+          urn: 'crn:contentful:::content:spaces/6fqi4ljzyr0e/environments/master/entries/9mpxT4zsRi6Iwukey8KeM',
+          type: 'ResourceLink',
+          linkType: 'Contentful:Entry',
+        },
+      },
+    };
+    const document: Document = inlineEntity(entry, INLINES.RESOURCE_HYPERLINK);
+    const expected = `<p><span>type: ${INLINES.RESOURCE_HYPERLINK} urn: ${entry.target.sys.urn}</span></p>`;
+
+    expect(documentToHtmlString(document)).toEqual(expected);
+  });
+
   it('renders embedded entry', () => {
     const entry = {
       target: {
         sys: {
           id: '9mpxT4zsRi6Iwukey8KeM',
-          link: 'Link',
-          type: 'Entry',
+          type: 'Link',
+          linkType: 'Entry',
         },
       },
     };
     const document: Document = inlineEntity(entry, INLINES.EMBEDDED_ENTRY);
     const expected = `<p><span>type: ${INLINES.EMBEDDED_ENTRY} id: ${entry.target.sys.id}</span></p>`;
+
+    expect(documentToHtmlString(document)).toEqual(expected);
+  });
+
+  it('renders embedded resource', () => {
+    const entry = {
+      target: {
+        sys: {
+          urn: 'crn:contentful:::content:spaces/6fqi4ljzyr0e/environments/master/entries/9mpxT4zsRi6Iwukey8KeM',
+          type: 'Link',
+          linkType: 'Contentful:Entry',
+        },
+      },
+    };
+    const document: Document = inlineEntity(entry, INLINES.EMBEDDED_RESOURCE);
+    const expected = `<p><span>type: ${INLINES.EMBEDDED_RESOURCE} urn: ${entry.target.sys.urn}</span></p>`;
 
     expect(documentToHtmlString(document)).toEqual(expected);
   });
@@ -316,5 +379,86 @@ describe('documentToHtmlString', () => {
 
   it('does not crash with undefined documents', () => {
     expect(documentToHtmlString(undefined as Document)).toEqual('');
+  });
+
+  it('preserves whitespace with preserveWhitespace option', () => {
+    const document: Document = {
+      nodeType: BLOCKS.DOCUMENT,
+      data: {},
+      content: [
+        {
+          nodeType: BLOCKS.PARAGRAPH,
+          data: {},
+          content: [
+            {
+              nodeType: 'text',
+              value: 'hello    world',
+              marks: [],
+              data: {},
+            },
+          ],
+        },
+      ],
+    };
+    const options: Options = {
+      preserveWhitespace: true,
+    };
+    const expected = '<p>hello&nbsp;&nbsp;&nbsp;&nbsp;world</p>';
+
+    expect(documentToHtmlString(document, options)).toEqual(expected);
+  });
+
+  it('preserves line breaks with preserveWhitespace option', () => {
+    const document: Document = {
+      nodeType: BLOCKS.DOCUMENT,
+      data: {},
+      content: [
+        {
+          nodeType: BLOCKS.PARAGRAPH,
+          data: {},
+          content: [
+            {
+              nodeType: 'text',
+              value: 'hello\nworld',
+              marks: [],
+              data: {},
+            },
+          ],
+        },
+      ],
+    };
+    const options: Options = {
+      preserveWhitespace: true,
+    };
+    const expected = '<p>hello<br/>world</p>';
+
+    expect(documentToHtmlString(document, options)).toEqual(expected);
+  });
+
+  it('preserves both spaces and line breaks with preserveWhitespace option', () => {
+    const document: Document = {
+      nodeType: BLOCKS.DOCUMENT,
+      data: {},
+      content: [
+        {
+          nodeType: BLOCKS.PARAGRAPH,
+          data: {},
+          content: [
+            {
+              nodeType: 'text',
+              value: 'hello   \n  world',
+              marks: [],
+              data: {},
+            },
+          ],
+        },
+      ],
+    };
+    const options: Options = {
+      preserveWhitespace: true,
+    };
+    const expected = '<p>hello&nbsp;&nbsp;&nbsp;<br/>&nbsp;&nbsp;world</p>';
+
+    expect(documentToHtmlString(document, options)).toEqual(expected);
   });
 });